Add files via upload
Uaktualniona wersja zadania 1 z listy 5

+++ PlayerAgeComparator.java
@@ -1,13 +1,11 @@
 package Zad1Lista5;
 
-import Zad1Lista5.Tenisista;
-
 import java.util.Comparator;
 
 public class PlayerAgeComparator implements Comparator<Tenisista> {
 
     @Override
     public int compare(Tenisista o1, Tenisista o2) {
-        return Double.compare(o1.getAge(),o2.getAge());
+        return Integer.compare(o1.getAge(),o2.getAge());
     }
 }
