+++ PlayerPointsComparator.java
@@ -5,6 +5,6 @@ import java.util.Comparator;
 public class PlayerPointsComparator implements Comparator<Tenisista> {
     @Override
     public int compare(Tenisista o1, Tenisista o2) {
-        return Double.compare(o2.getPoints(), o1.getPoints());
+        return Integer.compare(o2.getPoints(), o1.getPoints());
     }
 }
